Making file_writer lazily initialized (#358)
The `config.results_directory` string is being passed into the `FileWriter`
initializer. The `results_directory` can be changed via the `Allure.configure`
block. Since ruby passes strings by value, this necessitates that the
`Allure.configure` block be called prior to initializing the
`AllureLifecycle` object.

Moving the `FileWriter` to an instance method and memoizing such that it
will be initialized right before it is used for the first time. Thus
allowing for a less strict ordering.